+++ lib/widgets/images/svg_image.dart
@@ -1,37 +1,36 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
 import '../../functions/design.dart';
 
-class UdAssetImage extends StatelessWidget {
+class UdSvgImage extends StatelessWidget {
   final double height;
   final double width;
   final String imagePath;
   final BoxFit boxFit;
   final Alignment imageAlignment;
   final double borderRadius;
-  final Widget child;
 
-  /// Using network images are now more easy with UdNetworkImage. Just check parameters and pass values.
-  UdAssetImage({
+  /// Using svg images are now more easy with UdSvgImage. Just check parameters and pass values.
+  UdSvgImage({
     this.height,
     this.width,
     this.imagePath,
     this.boxFit,
     this.imageAlignment,
     this.borderRadius,
-    this.child,
   });
   @override
   Widget build(BuildContext context) {
     double _design = doNotUseThisDesignValue(context: context);
 
     if (imagePath == null) {
-      print("==UD==> Please provide [ imagePath ] for [ UdAssetImage ]");
+      print("==UD==> Please provide [ imagePath ] for [ UdSvgImage ]");
     }
     if (width == null) {
-      print("==UD==> Please provide [ width ] for [ UdAssetImage ]");
+      print("==UD==> Please provide [ width ] for [ UdSvgImage ]");
     }
     if (height == null) {
-      print("==UD==> Please provide [ height ] for [ UdAssetImage ]");
+      print("==UD==> Please provide [ height ] for [ UdSvgImage ]");
     }
     return imagePath != null && imagePath != ""
         ? Container(
@@ -41,13 +40,14 @@ class UdAssetImage extends StatelessWidget {
               borderRadius: BorderRadius.all(
                 Radius.circular(borderRadius ?? _design * 2),
               ),
-              image: DecorationImage(
-                alignment: imageAlignment ?? Alignment.center,
-                fit: boxFit ?? BoxFit.contain,
-                image: AssetImage(imagePath),
-              ),
             ),
-            child: child ?? SizedBox.shrink(),
+            child: SvgPicture.asset(
+              imagePath,
+              fit: boxFit ?? BoxFit.contain,
+              alignment: imageAlignment ?? Alignment.center,
+              height: height ?? _design * 50,
+              width: width ?? _design * 50,
+            ),
           )
         : SizedBox.shrink();
   }